Adjusted the concensus maker, need to build R file for postprocssing.

+++ Pipeline/workflow/processing.smk
@@ -1,16 +1,13 @@
 import subprocess
+import random
+import re
 
 configfile: "config.yaml"
 
-rule newall:
-    input:
-        #expand("Geneious/chimeras/single/{sample}_concensus.fasta", sample=config['single']),
-        expand("Postprocess/chimeras/paired/{sample}_concensus.fasta", sample=config['paired'])
-
 rule Bowtie_index_single:
-    priority: 10
+    priority: 5
     input:
-        "contigs/single/{sample}.Contigs.fasta",
+        "blastoutput/fetched/single/{sample}_accession.fasta",
     output:
         "Geneious/Bowtie2/single/{sample}/btbuild.log",
     log:
@@ -18,14 +15,19 @@ rule Bowtie_index_single:
     benchmark:
         "Geneious/Bowtie2/single/{sample}/benchmark/Bowtiebench.csv"
     shell:
-        """(
-        bowtie2-build {input} Geneious/Bowtie2/single/{wildcards.sample}/ref_genome_btindex > {output}
-        ) >{log} 2>&1"""
+        """
+        if [ -s {input[0]} ]; then
+            (bowtie2-build {input} Geneious/Bowtie2/single/{wildcards.sample}/ref_genome_btindex > {output}) >{log} 2>&1
+            echo {wildcards}
+        else
+            touch {output}
+        fi
+        """
 
 rule Bowtie_index_paired:
-    priority: 10
+    priority: 5
     input:
-        "contigs/paired/{sample}.Contigs.fasta",
+        "blastoutput/fetched/paired/{sample}_accession.fasta",
     output:
         "Geneious/Bowtie2/paired/{sample}/btbuild.log",
     log:
@@ -34,12 +36,16 @@ rule Bowtie_index_paired:
         "Geneious/Bowtie2/paired/{sample}/benchmark/Bowtiebench.csv"
     shell:
         """
-        (
-        bowtie2-build {input} Geneious/Bowtie2/paired/{wildcards.sample}/ref_genome_btindex > {output}
-        ) >{log} 2>&1"""
+        if [ -s {input[0]} ]; then
+            (bowtie2-build {input} Geneious/Bowtie2/paired/{wildcards.sample}/ref_genome_btindex > {output}
+            ) >{log} 2>&1
+        else
+            touch {output}
+        fi
+        """
 
 rule Bowtie2_single:
-    priority: 7
+    priority: 4
     input:
         "contigs/single/{sample}.Contigs.fasta",
         "Geneious/Bowtie2/single/{sample}/btbuild.log"
@@ -55,11 +61,15 @@ rule Bowtie2_single:
         "Geneious/Bowtie2/single/benchmark/{sample}_bench.txt"
     shell:
         """
-        bowtie2 -x {params.btindex} -fa {input} -S {output} 2>> {log} && touch {output}
+        if [ -s {input[0]} ]; then
+            bowtie2 -x {params.btindex} -fa {input[0]} -S {output} 2>> {log} && touch {output}
+        else
+            touch {output}
+        fi
         """
 
 rule Bowtie2_paired:
-    priority: 7
+    priority: 4
     input:
         "contigs/paired/{sample}.Contigs.fasta",
         "Geneious/Bowtie2/paired/{sample}/btbuild.log"
@@ -75,23 +85,85 @@ rule Bowtie2_paired:
         "Geneious/Bowtie2/paired/benchmark/{sample}_bench.txt"
     shell:
         """
-        bowtie2 -x {params.btindex} -fa {input[0]} -S {output} 2>> {log} && touch {output}
+        if [ -s {input[0]} ]; then
+            bowtie2 -x {params.btindex} -fa {input[0]} -S {output} 2>> {log} && touch {output}
+        else
+            touch {output}
+        fi
         """
+
+rule adjust_sam_paired:
+    priority: 3
+    input:
+        "Geneious/Bowtie2/paired/mapped/{sample}.sam"
+    output:
+        "Geneious/Bowtie2/paired/done/{sample}.sam"
+    run:
+        subprocess.call(f"touch {output}", shell=True)
+        myfile = open(f'{input}')
+        contigs = []
+        with open(f'{output}',"r+") as newtest:
+            for line in myfile:
+                mycontig = ""
+                if line.startswith("@SQ"):
+                    contig = (re.findall("[a-z]*-[0-9]*",line))
+                    if contig not in contigs:
+                        contigs.append(contig)
+                        newtest.write(line)
+                    else:
+                        for item in contig:
+                            newtest.write(re.sub("[a-z]*-[0-9]*","".join(
+                                str(item).split("-")[0]) + "-" + str(random.randint(0,10000)),line,count=0,flags=0))
+                else:
+                    newtest.write(line)
+
+rule adjust_sam_single:
+    priority: 3
+    input:
+        "Geneious/Bowtie2/single/mapped/{sample}.sam"
+    output:
+        "Geneious/Bowtie2/single/done/{sample}.sam"
+    run:
+        subprocess.call(f"touch {output}", shell=True)
+        myfile = open(f'{input}')
+        contigs = []
+        with open(f'{output}',"r+") as newtest:
+            for line in myfile:
+                mycontig = ""
+                if line.startswith("@SQ"):
+                    contig = (re.findall("[a-z]*-[0-9]*",line))
+                    if contig not in contigs:
+                        contigs.append(contig)
+                        newtest.write(line)
+                    else:
+                        for item in contig:
+                            newtest.write(re.sub("[a-z]*-[0-9]*","".join(
+                                str(item).split("-")[0]) + "-" + str(random.randint(0,10000)),line,count=0,flags=0))
+                else:
+                    newtest.write(line)
 
 rule create_consencus_paired:
     priority: 4
     input:
         "blastoutput/fetched/paired/{sample}_accession.fasta",
-        "Geneious/Bowtie2/paired/mapped/{sample}.sam"
+        "Geneious/Bowtie2/paired/done/{sample}.sam"
     output:
         "Geneious/chimeras/paired/{sample}_concensus.fasta"
+    log:
+        "Geneious/chimeras/paired/{sample}.log"
     shell:
         """
-        samtools view -bS {input[1]}| samtools sort -o Geneious/chimeras/paired/{wildcards.sample}_bowtie.bam
-        bcftools mpileup -Ou -f {input[0]} Geneious/chimeras/paired/{wildcards.sample}_bowtie.bam | bcftools call -mv -Oz -o Geneious/chimeras/paired/{wildcards.sample}_calls.vcf.gz && bcftools index Geneious/chimeras/paired/{wildcards.sample}_calls.vcf.gz
-        bcftools norm -f {input[0]} Geneious/chimeras/paired/{wildcards.sample}_calls.vcf.gz -Ob -o Geneious/chimeras/paired/{wildcards.sample}_calls.norm.bcf
-        bcftools filter --IndelGap 5 Geneious/chimeras/paired/{wildcards.sample}_calls.norm.bcf -Ob -o Geneious/chimeras/paired/{wildcards.sample}_calls.norm.flt-indels.bcf
-        cat {input[0]} | bcftools consensus Geneious/chimeras/paired/{wildcards.sample}_calls.vcf.gz > {output} && touch {output}
+        if [ -s {input[0]} ]; then
+            (
+            samtools view -bS {input[1]}| samtools sort -o Geneious/chimeras/paired/{wildcards.sample}_bowtie.bam
+            bcftools mpileup -Ou -f {input[0]} Geneious/chimeras/paired/{wildcards.sample}_bowtie.bam | bcftools call -mv -Oz -o Geneious/chimeras/paired/{wildcards.sample}_calls.vcf.gz && bcftools index Geneious/chimeras/paired/{wildcards.sample}_calls.vcf.gz
+            bcftools norm -f {input[0]} Geneious/chimeras/paired/{wildcards.sample}_calls.vcf.gz -Ob -o Geneious/chimeras/paired/{wildcards.sample}_calls.norm.bcf
+            bcftools filter --IndelGap 5 Geneious/chimeras/paired/{wildcards.sample}_calls.norm.bcf -Ob -o Geneious/chimeras/paired/{wildcards.sample}_calls.norm.flt-indels.bcf
+            cat {input[0]} | bcftools consensus Geneious/chimeras/paired/{wildcards.sample}_calls.vcf.gz > {output} && touch {output}
+            ) >{log} 2>&1
+        else
+            touch {output}
+        fi
         """
 
 
@@ -99,16 +171,20 @@ rule create_concensus_single:
     priority: 4
     input:
         "blastoutput/fetched/single/{sample}_accession.fasta",
-        "Geneious/Bowtie2/single/mapped/{sample}.sam"
+        "Geneious/Bowtie2/single/done/{sample}.sam"
     output:
         "Geneious/chimeras/single/{sample}_concensus.fasta"
     shell:
         """
-        samtools view -bS {input[1]}| samtools sort -o Geneious/chimeras/single/{wildcards.sample}_bowtie.bam
-        bcftools mpileup -Ou -f {input[0]} Geneious/chimeras/single/{wildcards.sample}_bowtie.bam | bcftools call -mv -Oz -o Geneious/chimeras/single/{wildcards.sample}_calls.vcf.gz && bcftools index Geneious/chimeras/single/{wildcards.sample}_calls.vcf.gz
-        bcftools norm -f {input[0]} Geneious/chimeras/single/{wildcards.sample}_calls.vcf.gz -Ob -o Geneious/chimeras/single/{wildcards.sample}_calls.norm.bcf
-        bcftools filter --IndelGap 5 Geneious/chimeras/single/{wildcards.sample}_calls.norm.bcf -Ob -o Geneious/chimeras/single/{wildcards.sample}_calls.norm.flt-indels.bcf
-        cat {input[0]} | bcftools consensus Geneious/chimeras/single/{wildcards.sample}_calls.vcf.gz > {output} && touch {output}
+        if [ -s {input[0]} ]; then
+            samtools view -bS {input[1]}| samtools sort -o Geneious/chimeras/single/{wildcards.sample}_bowtie.bam
+            bcftools mpileup -Ou -f {input[0]} Geneious/chimeras/single/{wildcards.sample}_bowtie.bam | bcftools call -mv -Oz -o Geneious/chimeras/single/{wildcards.sample}_calls.vcf.gz && bcftools index Geneious/chimeras/single/{wildcards.sample}_calls.vcf.gz
+            bcftools norm -f {input[0]} Geneious/chimeras/single/{wildcards.sample}_calls.vcf.gz -Ob -o Geneious/chimeras/single/{wildcards.sample}_calls.norm.bcf
+            bcftools filter --IndelGap 5 Geneious/chimeras/single/{wildcards.sample}_calls.norm.bcf -Ob -o Geneious/chimeras/single/{wildcards.sample}_calls.norm.flt-indels.bcf
+            cat {input[0]} | bcftools consensus Geneious/chimeras/single/{wildcards.sample}_calls.vcf.gz > {output} && touch {output}
+        else
+            touch {output}
+        fi
         """
 
 rule adjust_concensus_paired:
@@ -142,6 +218,7 @@ rule adjust_concensus_single:
                     newtest.write("".join(f">{wildcards.sample}" + "_" + line.strip(">")))
                 else:
                     newtest.write(line)
+
 
 onsuccess:
     with open("process_dag.txt","w") as f:
